requirements.txt
Flask==3.1.2
firebase-admin==6.5.0
requests==2.32.3
python-dotenv==1.0.1
pytest-cov